Move auth header service to e2e folder
