--- test/auth-header/client.html
@@ -1,67 +1,0 @@
-<html>
-<head>
-  <style>
-    iframe {
-      width: 100%;
-      height: 300px;
-    }
-  </style>
-</head>
-<body>
-
-  <section>
-    <h2>Response for same-origin:</h2>
-    <pre id="responseSameOriginEl"></pre>
-  <section>
-
-  <section>
-    <h2>Response for x-origin:</h2>
-    <pre id="responseCrossOriginEl"></pre>
-  <section>
-
-<script>
-
-function setCookie() {
-  return Promise.all([
-    fetch('//localhost:8000/test/auth-header/service/set-cookie', {
-      credentials: 'include',
-    }),
-    fetch('//pub.localhost:8000/test/auth-header/service/set-cookie', {
-      credentials: 'include',
-    })
-  ]);
-}
-
-function runTests() {
-  fetch('//localhost:8000/test/auth-header/service/echo', {
-    method: 'post',
-    credentials: 'omit',
-    headers: {
-      'Authorization': 'Bearer 12345',
-    },
-  }).then(response => response.json()).then(json => {
-    console.log('Response: ', json);
-    responseSameOriginEl.textContent = JSON.stringify(json, true, 2);
-  }).catch(reason => {
-    responseSameOriginEl.textContent = 'Error: ' + reason;
-  });
-
-  fetch('//pub.localhost:8000/test/auth-header/service/echo', {
-    method: 'post',
-    credentials: 'omit',
-    headers: {
-      // Authorization always requires pre-flight
-    },
-  }).then(response => response.json()).then(json => {
-    console.log('Response: ', json);
-    responseCrossOriginEl.textContent = JSON.stringify(json, true, 2);
-  }).catch(reason => {
-    responseCrossOriginEl.textContent = 'Error: ' + reason;
-  });
-}
-
-setCookie().then(runTests);
-
-</script>
-</body>
-</html>
